requirements.txt
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.10" and python_full_version < "3.11.3"
cffi==1.17.1 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
click==8.1.8 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
contextlib2==21.6.0 ; python_version >= "3.10" and python_version < "4.0"
cryptography==44.0.0 ; python_version >= "3.10" and python_version < "4.0"
django-cors-headers==4.6.0 ; python_version >= "3.10" and python_version < "4.0"
django-dbbackup==4.2.1 ; python_version >= "3.10" and python_version < "4.0"
django-filter==24.3 ; python_version >= "3.10" and python_version < "4.0"
django-ninja-extra==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
django-ninja-jwt==5.3.5 ; python_version >= "3.10" and python_version < "4.0"
django-ninja==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
django-redis==5.4.0 ; python_version >= "3.10" and python_version < "4.0"
django-rq==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
django-sendfile2==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
django==5.1.4 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework-simplejwt==5.4.0 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.15.2 ; python_version >= "3.10" and python_version < "4.0"
drf-chunked-upload==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.10" and python_version < "4.0"
hiredis==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
injector==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
markdown==3.7 ; python_version >= "3.10" and python_version < "4.0"
numpy==2.2.1 ; python_version >= "3.10" and python_version <= "3.11" or python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
psycopg2==2.9.10 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.27.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.10.5 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
pyjwt[crypto]==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.10" and python_version < "4.0"
redis==5.2.1 ; python_version >= "3.10" and python_version < "4.0"
rq==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.10" and python_version < "4.0"
whitenoise==6.8.2 ; python_version >= "3.10" and python_version < "4.0"
